Implemented parsing of scalars and complex types
This still isn't sufficient. We need to marry the document and complex-
type namespaces so that we can have a unified lookup for any value
described in a namespace. This is made simpler by the fact that the Go
XML parser takes care of itself in a correct, namespace-annotated
document. We just need them when it comes to looking up values for
formal XML 2-tuple names.