Begin implementation of array semantics
Right now, we're identifying that we have char-data and that the current
node is an "li", we then match for a known array-type parent path
pattern. We'd like to refactor to support array-types in the namespace
definitions so that we can process this, downward, *from* a node that is
known to contain an array type rather than searching upward from
something that *looks like* an array. This latter method is not as tight
and likely to waste cycles.